edit capture source of capture.cgi
