+++ .gitignore
@@ -1,2 +1,4 @@
 **/build/
 .vscode/
+*.mov
+*.mp4
